+++ public/assets/json/filtros.json
@@ -622,7 +622,7 @@
                         "modelo": "Seccionales"
                     },
                     {
-                        "modelo": "Sófas camas"
+                        "modelo": "Sófa cama"
                     },
                     {
                         "modelo": "Reclinables"
