+++ public/assets/json/ofertas.json
@@ -3,5 +3,6 @@
     "K231113N11", "K2241113N23", "K2242113N10", "K2242113N12", "K54224N10", "K54224N11", "K52130N10", "K52221N14", "K52241N10",
     "K1411MFN10", "K148N10", "K324235N41", "K324236N45", "K324253N2O", "K3242113N10", "K3232116N12", "K324233N10", "K324236N47",
     "K324233N30", "K3215232N15", "K323236N98", "K324273N12", "K77412N11", "K7342N13", "K323235N29", "K334153N13", "K324235N43", "K324235N11",
-    "K43212N17", "K4131N10", "K41212N18", "K41432N23", "K43212N18", "K43212N12", "K43212N32", "K641N10", "K6222N16"
+    "K43212N17", "K4131N10", "K41212N18", "K41432N23", "K43212N18", "K43212N12", "K43212N32", "K641N10", "K6222N16", "K3215232N18", "K322232N10",
+    "K322232N18", "K322232N15", "K322232N13"
 ]
