+++ src/Componentes/Elementos/Spinner/Spinner.css
@@ -16,4 +16,4 @@
     100%{
         transform: rotate(360deg);
     }
-} 
+}
