+++ src/Componentes/Filtros/Filtros.css
@@ -4,7 +4,6 @@
     display: flex;
     flex-direction: column;
     gap: 20px;
-    /* border: 1px solid #D9D9D9; */
     background: var(--white);
     border-radius: 6px;
 }
@@ -38,7 +37,7 @@
 }
 
 .filtros-content-categorias{
-    padding: 20px 0;
+    padding: 10px 0;
 }
 
 .filtros-content-categorias .filtro-items{
@@ -61,6 +60,10 @@
     gap: 20px;
 }
 
+.filtro-detalle-title{
+    text-transform: lowercase !important;
+}
+
 .filtros-content > div:nth-child(1) > ul{
     height: auto !important;
     padding: 10px 0 !important;
@@ -70,7 +73,7 @@
 .filtro-items{
     display: flex;
     width: 100%;
-    padding: 10px 0;
+    padding-top: 10px;
     flex-wrap: wrap;
     gap: 5px;
     overflow: hidden;
