+++ src/Componentes/Plantillas/Producto/Producto.css
@@ -15,7 +15,7 @@
     flex-direction: column;
     width: 100%;
     height: 100%;
-    min-height: 380px;
+    min-height: 320px;
     background: var(--producto);
     overflow: hidden !important;
 }
@@ -187,9 +187,13 @@
 }
 
 .product-card-separar{
+    position: absolute;
+    left: 10px;
+    bottom: 100%;
     display: flex;
     align-items: center;
     justify-content: center;
+    z-index: 15;
     gap: 3px;
 }
