--- src/Paginas/Categorias/Componentes/Filtros/Filtros.css
@@ -1,348 +1,0 @@
-.filters-content{
-    width: 100%;
-    min-height: 600px;
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    padding: 20px;
-    background: var(--white);
-    border-radius: 6px;
-}
-
-.filters-content-top{
-    display: none;
-}
-
-.price-range{
-    display: flex;
-    flex-direction: column;
-}
-
-.price-range .filter-title{
-    margin-bottom: 20px;
-}
-
-.price-range ul{
-    gap: 5px;
-}
-
-.price-range ul li{
-    display: flex;
-    align-items: center;
-    justify-content: left;
-    gap: 5px;
-}
-
-.radio-input{
-    display: none;
-}
-
-.radio-label{
-    position: relative;
-    padding-left: 20px;
-    cursor: pointer;
-    user-select: none;
-}
-
-.radio-label::before{
-    content: "";
-    position: absolute;
-    left: 0;
-    top: 50%;
-    translate: 0 -50%;
-    width: 10px;
-    height: 10px;
-    border: 2px solid var(--white);
-    border-radius: 100%;
-    background: var(--color-1);
-    transition: var(--t-200ms-linear) !important;
-    z-index: 1;
-}
-
-.radio-input:checked + .radio-label::after{
-    content: "";
-    position: absolute;
-    left: -3px;
-    top: 50%;
-    translate: 0 -50%;
-    width: 20px;
-    height: 20px;
-    border-radius: 100%;
-    background: var(--color-1);
-    opacity: 0;
-    transition: var(--t-200ms-linear) !important;
-    z-index: 0;
-}
-
-.radio-input:checked + .radio-label::after{
-    opacity: 1;
-}
-
-.filter-title{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.filter-title span{
-    display: none;
-}
-
-.filtro-envio-gratis{
-    width: 100%;
-    height: 60px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 10px;
-    border-top: 1px solid rgba(0 0 0 / 0.25);
-    border-bottom: 1px solid rgba(0 0 0 / 0.25);
-}
-
-.filtro-envio-gratis > div span,
-.filtro-envio-gratis > div p{
-    color: #37c937;
-}
-
-.filtro-envio-gratis button{
-    width: 36px;
-    height: 20px;
-    background: #999;
-    border-radius: 25px;
-    position: relative;
-    transition: var(--t-200ms-linear);
-}
-
-.filtro-envio-gratis button.active{
-    background: #37c937;
-}
-
-.filtro-envio-gratis button span{
-    position: absolute;
-    top: 2px;
-    left: 2px;
-    width: auto;
-    height: calc(100% - 4px);
-    aspect-ratio: 1/1;
-    background: white;
-    border-radius: 100%;
-    transition: var(--t-200ms-linear);
-}
-
-.filtro-envio-gratis button.active span{
-    left: calc(100% - 18px);
-}
-
-.filter ul{
-    display: flex;
-    flex-wrap: wrap;
-    gap: 5px;
-}
-
-.filter ul li{
-    position: relative;
-    cursor: pointer;
-    border-radius: 3px;
-    background: rgba(0 0 0 / 0.05);
-    border: 1px solid rgba(0 0 0 / 0.3);
-    padding: 2px 9px;
-    overflow: hidden;
-    transition: var(--t-200ms-linear);
-}
-
-.filter ul li:hover{
-    border: 1px solid var(--color-2);
-    background: var(--color-2);
-}
-
-.filter ul li.active{
-    border: 1px solid var(--color-1);
-    background: var(--color-1);
-}
-
-.filter ul li input{
-    cursor: pointer;
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    opacity: 0;
-}
-
-.filter ul li label{
-    font-size: 13px;
-    color: var(--text);
-}
-
-.filter ul li:hover label,
-.filter ul li.active label{
-    color: var(--white);
-    transition: var(--t-200ms-linear);
-}
-
-.filters-clear-container{
-    width: 100%;
-    display: flex;
-    /* margin-top: auto; */
-}
-
-.filters-clear{
-    margin: auto;
-    width: 100%;
-    height: 48px;
-    background: var(--color-2);
-    transition: var(--t-200ms-linear);
-    border-radius: 6px;
-}
-
-.filters-clear:hover{
-    background: var(--color-1);
-}
-
-.filters-clear p,
-.filters-clear span{
-    color: var(--white);
-}
-
-.filters-layer{
-    display: none;
-}
-
-@media screen and (max-width: 600px){
-    .filters-layer{
-        display: flex;
-        position: fixed;
-        left: 0;
-        top: 0;
-        width: 100%;
-        height: 100vh;
-        background: rgba(0 0 0 / 0.25);
-        transition: var(--t-200ms-linear);
-        z-index: 149;
-
-        opacity: 0;
-        pointer-events: none;
-        user-select: none;
-    }
-
-    .filters-layer.active{
-        opacity: 1;
-        pointer-events: unset;
-        user-select: unset;
-    }
-
-    .filters-container{
-        position: fixed;
-        left: 100%;
-        top: 0;
-        width: 70%;
-        height: 100vh;
-        background: var(--white);
-        z-index: 150;
-        padding: 20px 0;
-        overflow-y: auto;
-        border-left: 1px solid rgba(0 0 0 / 0.25);
-        transition: var(--t-200ms-linear);
-    }
-
-    .filters-container.active{
-        left: 30%;
-    }
-
-    .filters-content{
-        width: 100%;
-        height: auto;
-        padding: 0;
-        gap: 0;
-    }
-
-    .filters-content-top{
-        width: 100%;
-        padding: 0 20px;
-        margin-bottom: 20px;
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-    }
-
-    .close-filters{
-        display: flex;
-    }
-
-    .close-filters span{
-        color: var(--color-1);
-    }
-
-    .price-range .filter-title{
-        margin-bottom: 0px;
-    }
-
-    .price-range .filter-title:hover{
-        background: unset !important;
-    }
-
-    .price-range > ul{
-        padding: 10px 20px !important;
-        height: auto !important;
-    }
-
-    .filtro-envio-gratis{
-        padding: 0 20px;
-    }
-
-    .filtro-envio-gratis button{
-        width: 40px;
-        height: 24px;
-    }
-
-    .filtro-envio-gratis button.active span{
-        left: calc(100% - 22px);
-    }
-
-    .filter{
-        gap: 0 !important;
-        border-bottom: 1px solid rgba(0 0 0 / 0.25);
-    }
-
-    .filter-title{
-        height: 48px;
-        padding: 0 20px;
-        cursor: pointer;
-        transition: var(--t-200ms-linear);
-    }
-
-    .filter-title:hover{
-        background: rgba(0 0 0 / 0.05);
-    }
-
-    .filter-title > p{
-        font-size: 15px;
-    }
-
-    .filter-title span{
-        display: flex;
-        color: var(--text);
-    }
-
-    .filters-content > div > ul{
-        overflow: hidden;
-        height: 0;
-        padding: 0 20px;
-        transition: var(--t-200ms-linear);
-        interpolate-size: allow-keywords;
-    }
-
-    .filters-content > div > ul.active{
-        padding: 10px 20px;
-        height: auto;
-    }
-
-    .filters-clear-container{
-        margin-top: 20px;
-    }
-
-    .filters-clear{
-        width: calc(100% - 40px);
-    }
-}
