+++ src/Paginas/PaginaPrincipal/Componentes/Ofertas/Ofertas.css
@@ -1,44 +1,270 @@
-.homepage-offers-container{
-    overflow: hidden;
-}
-
-.homepage-ofertas-content{
-    position: relative;
-}
-
-.homepage-offers-content{
-    display: grid;
-    grid-template-columns: repeat(6, 1fr);
-    gap: 10px;
-}
-
-.homepage-offers-content li a img{
+.block-content-solo-por-horas{
+    background: #222;
+    padding: 20px;
     border-radius: 6px;
 }
 
-/**/
+.block-content-solo-por-horas .block-title-container{
+    width: 100%;
+    margin-right: auto;
+    display: flex !important;
+    align-items: center;
+    justify-content: space-between !important;
+    position: relative;
+    margin-bottom: 0 !important;
+}
 
-.homepage-ofertas-button{
+.block-content-solo-por-horas .block-title-container .block-title{
+    width: auto !important;
+    color: var(--white);
+}
+
+.block-content-solo-por-horas .block-title-container .sale-time{
+    margin-left: auto;
+}
+
+.solo-por-horas-productos{
+    width: 100%;
+    display: grid;
+    grid-template-columns: repeat(5, 1fr);
+    grid-template-rows: repeat(2, 1fr);
+    gap: 10px;
+}
+
+.solo-por-horas-productos > li{
+    width: 100%;
+    height: 100%;
+    display: flex;
+    background: var(--white);
+    border-radius: 6px;
+    overflow: hidden;
+}
+
+.solo-por-horas-productos > li:nth-child(1){
+    grid-column: 1/2;
+    grid-row: 1/-1;
+}
+
+.solo-por-horas-productos > li:nth-child(4){
+    grid-column: 3/4;
+    grid-row: 1/-1;
+}
+
+.solo-por-horas-productos > li:nth-child(7){
+    grid-column: 5/6;
+    grid-row: 1/-1;
+}
+
+.solo-por-horas-productos > li .product-card{
+    background: var(--white);
+    width: 100%;
+    height: 100%;
+    min-height: unset !important;
+    display: flex;
+    flex-direction: column;
+    border-radius: 0 !important;
+    border: 0 !important;
+}
+
+.solo-por-horas-productos > li .product-card:hover{
+    border: 0 !important;
+}
+
+.solo-por-horas-productos > li .product-card .product-card-image img{
+    width: 100%;
+}
+
+.product-card-image-container{
+    width: 100%;
+    height: 200px;
+    position: relative;
+}
+
+.product-card-image-container .descuento{
     position: absolute;
-    bottom: 196px;
+    left: 10px;
+    top: 10px;
     display: flex;
     align-items: center;
     justify-content: center;
-    width: 48px;
-    height: 48px;
-    overflow: hidden;
-    border-radius: 100%;
-    background: var(--white);
+    font-weight: bold;
+    color: crimson;
+    font-size: 24px;
 }
 
-.homepage-ofertas-button-1{
-    left: 10px;
-}
-
-.homepage-ofertas-button-2{
+.product-card-image-container .tipo-de-envio{
+    position: absolute;
     right: 10px;
+    bottom: 10px;
+    background: orange;
+    padding: 4px 8px !important;
+    border: 0 !important;
+    font-weight: bold;
+    background: #dcfcdc;
+    font-size: 14px;
+    color: green;
+    border-radius: 100px !important;
 }
 
-.homepage-ofertas-button span{
-    color: var(--color-1);
+.product-card-image-container .tipo-de-envio:hover{
+    background: 0 !important;
+    border: 0 !important;
+}
+
+.product-card-image-slider{
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+}
+
+.product-card-image-slides{
+    width: 200%;
+    height: 100%;
+    display: flex;
+}
+
+.product-card-image-slides li{
+    width: 100%;
+    height: 100%;
+    display: flex;
+}
+
+.product-card-image-slides li img{
+    margin: auto;
+    display: flex;
+    width: 80%;
+    max-width: 180px;
+    aspect-ratio: 1/1;
+}
+
+.block-content-solo-por-horas > ul li .product-card > div:nth-child(2){
+    width: 100%;
+    height: calc(100% - 200px);
+    padding: 0 10px 10px 10px;
+}
+
+.block-content-solo-por-horas > ul li .product-card > div:nth-child(2) p.title{
+    font-size: 16px;
+    color: var(--text);
+}
+
+.block-content-solo-por-horas .product-card .product-card-content .product-card-ofert,
+.block-content-solo-por-horas .product-card .product-card-content .product-card-tipo-de-envio{
+    position: unset;
+    margin-left: auto;
+}
+
+@media screen and (max-width: 800px){
+    .solo-por-horas-productos{
+        grid-template-columns: repeat(3, 1fr) !important;
+        grid-template-rows: repeat(4, auto) !important;
+    }
+
+    .solo-por-horas-productos > li:nth-child(1){
+        grid-column: 1/2;
+        grid-row: 1/3;
+    }
+
+    .solo-por-horas-productos > li:nth-child(2){
+        grid-column: 2/3;
+        grid-row: 1/2;
+    }
+
+    .solo-por-horas-productos > li:nth-child(3){
+        grid-column: 3/4;
+        grid-row: 3/4;
+    }
+
+    .solo-por-horas-productos > li:nth-child(4){
+        grid-column: 3/4;
+        grid-row: 1/3;
+    }
+
+    .solo-por-horas-productos > li:nth-child(5){
+        grid-column: 1/2;
+        grid-row: 3/4;
+    }
+
+    .solo-por-horas-productos > li:nth-child(6){
+        grid-column: 1/2;
+        grid-row: 4/5;
+    }
+
+    .solo-por-horas-productos > li:nth-child(7){
+        grid-column: 2/3;
+        grid-row: 2/4;
+    }
+}
+
+@media screen and (max-width: 600px){
+    .block-content-solo-por-horas .block-title-container{
+        flex-direction: column-reverse;
+        gap: 20px;
+    }
+
+    .solo-por-horas-productos{
+        grid-template-columns: repeat(2, 1fr) !important;
+        grid-template-rows: repeat(5, auto) !important;
+    }
+
+    .solo-por-horas-productos > li:nth-child(1){
+        grid-column: 1/-1;
+        grid-row: 1/2;
+    }
+
+    .solo-por-horas-productos > li:nth-child(4){
+        grid-column: 1/-1;
+        grid-row: 3/4;
+    }
+
+    .solo-por-horas-productos > li:nth-child(7){
+        grid-column: 1/-1;
+        grid-row: 5/6;
+    }
+
+    .solo-por-horas-productos > li:nth-child(2),
+    .solo-por-horas-productos > li:nth-child(3){
+        grid-row: 2/3 !important;
+    }
+
+    .solo-por-horas-productos > li:nth-child(2){
+        grid-column: 1/2;
+    }
+
+    .solo-por-horas-productos > li:nth-child(3){
+        grid-column: 2/3;
+    }
+
+    .solo-por-horas-productos > li:nth-child(5),
+    .solo-por-horas-productos > li:nth-child(6){
+        grid-row: 4/5 !important;
+    }
+
+    .solo-por-horas-productos > li:nth-child(5){
+        grid-column: 1/2;
+    }
+
+    .solo-por-horas-productos > li:nth-child(6){
+        grid-column: 2/3;
+    }
+
+    .solo-por-horas-productos > li:nth-child(1) .product-card,
+    .solo-por-horas-productos > li:nth-child(4) .product-card,
+    .solo-por-horas-productos > li:nth-child(7) .product-card{
+        display: grid !important;
+        grid-template-columns: repeat(2, 1fr) !important;
+    }
+
+    .solo-por-horas-productos > li:nth-child(1) .product-card .product-card-images,
+    .solo-por-horas-productos > li:nth-child(4) .product-card .product-card-images,
+    .solo-por-horas-productos > li:nth-child(7) .product-card .product-card-images{
+        height: 100% !important;
+    }
+
+    .solo-por-horas-productos > li:nth-child(1) .product-card .product-card-content,
+    .solo-por-horas-productos > li:nth-child(4) .product-card .product-card-content,
+    .solo-por-horas-productos > li:nth-child(7) .product-card .product-card-content{
+        height: 100%;
+        padding: 10px 10px 10px 0 !important;
+    }
 }
