+++ src/Paginas/PaginaPrincipal/Componentes/Perfiles/Perfiles.js
@@ -5,7 +5,7 @@ function Perfiles(){
         <section className='block-container'>
             <div className='block-content'>
                 <div className='d-flex-column gap-10'>
-                    <p className='title uppercase text'>Calidad y respaldo presentes en las principales tiendas del Perú</p>
+                    <p className='block-title text-left text'>Calidad y respaldo presentes en las principales tiendas del Perú</p>
                     <ul className='profiles'>
                         <li className='profile'>
                             <img src="/assets/imagenes/paginas/pagina-principal/perfiles/homesleep.webp" alt="Homesleep" className='profile-card-banner' />
