+++ src/Paginas/PaginaPrincipal/Componentes/Slider/Slider.css
@@ -10,15 +10,14 @@
     margin: 0 auto;
     width: auto;
     max-width: var(--content-mx-width);
-    height: auto;
+    height: 100%;
     max-height: 56px;
 }
 
 .hero-container{
     margin: auto;
     width: 100%;
-    height: calc(100vh - 165px);
-    max-height: 600px;
+    height: 100%;
     position: relative;
     overflow: hidden;
 }
